// src/features/voca/types/ExerciseType.ts
export enum ExerciseType {
  CHOOSE_WORD_BASED_ON_AUDIO = "CHOOSE_WORD_BASED_ON_AUDIO", // roi
  CHOOSE_WORD_BASED_ON_MEANING = "CHOOSE_WORD_BASED_ON_MEANING",
  CHOOSE_WORD_BASED_ON_DEFINITION = "CHOOSE_WORD_BASED_ON_DEFINITION",
  CHOOSE_WORD_BASED_ON_SUGGESTION = "CHOOSE_WORD_BASED_ON_SUGGESTION", // roi

  CHOOSE_MEANING_BASED_ON_WORD = "CHOOSE_MEANING_BASED_ON_WORD", // roi

  CHOOSE_DEFINITION_BASED_ON_WORD = "CHOOSE_DEFINITION_BASED_ON_WORD",

  FILL_WORD_BASED_ON_AUDIO = "FILL_WORD_BASED_ON_AUDIO",
  FILL_WORD_BASED_ON_DEFINITION = "FILL_WORD_BASED_ON_DEFINITION",
  FILL_WORD_BASED_ON_SUGGESTION = "FILL_WORD_BASED_ON_SUGGESTION",
}